% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Fatality Accident Reporting System--Mapping Function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numerical values which are integers that represents
a spefic state and the year is the year of interest in mapping the incidents.
state.num value in the FARS range from 1 to 56. Integer value outside this
range will result in an error.}

\item{year}{Numerical values and they must be four digit year integer value.}
}
\value{
function will return a map in a graphical form with incidents represented
 in a particular state based on locational information specified in FARS.
}
\description{
\code{fars_map_state(state.num, year)} takes in the data from the FARS files and maps the
 number of incidents for the specified state for the given year. This function uses
 make_filename and fars_read functions. It also requires *dplyr*, *maps*, and *graphics*
 packages to use filter, map, and points functions, respectively.
}
\examples{
\dontrun{
 fars_map_state(1, 2013)
 fars_map_state(20, 2014)
 fars_map_state(50, 2015)
 }

}
