% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{FARS (Fatality Accident Reporting System) Read File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{String characters used to read the file. This parameter takes the output of
make_filename function. Although the filename can be created as a string input
into the function as a parameter, any mis-spelled component or mis-specified
filename will result in an error.}
}
\value{
output of this function is a dataframe and S3 class tbl_df from
 dplyr package.
}
\description{
\code{fars_read(filename)} checks the existence of the file (based on the filename
 supplied in the parameter), reads the file, and loads the file as a data frame
 and S3 class "tbl_df" from dplyr package.
}
